chore: fix lint errors after rebase onto main
